requirements.txt
sounddevice
numpy
rich
requests
pycaw
comtypes
pywin32
psutil
python-dotenv